added dynamic subscriptions for events topic
Both agents of users and services must receive some events (for
example metric.pull).
This allows kruonis to respond to dynamic subscription requests
which is the only way for agents of users to subscribe to a topic of an
app.

This doesnt implement events separation into two visibility scopes:
external (which are intended for both services and users) and internal
(which are intended only for services).